+++ frontend/src/App.module.scss
@@ -17,23 +17,19 @@
 
 .slideRoot {
   padding-top: 100px;
+  width: 100%;
+  height: 100%;
 }
 
 .iphone {
   color: white;
   background: center / cover no-repeat url('./assets/iphone.png');
-  width: 100%;
-  height: 100%;
 }
 
 .tablet {
   background: center / cover no-repeat url('./assets/tablet.png');
-  width: 100%;
-  height: 100%;
 }
 
 .airpods {
   background: center / cover no-repeat url('./assets/airpods.png');
-  width: 100%;
-  height: 100%;
 }
